==> 01_CIC_Cleaning.ipynb <==
# Cleaning CIC IoT-DIAD Dataset

## Overview: Loading, Initial Observations, and Cleaning CIC IoT-DIAD 2024 dataset


---
## Project Workflow:
1. **Step 1: Install and Import Necessary Libraries**
2. **Step 2: Access Data**
3. **Step 3: Check Packaging**
5. **Step 4: Clean Data**



# Step 1: Install and Import Necessary Libraries


import pandas as pd
import numpy as np


# Step 2: Access Data

# brute force
file_path = r"data/DictionaryBruteForce.csv"

BF_data = pd.read_csv(file_path)

file_path = r"data/BenignTraffic.csv"

BN_data = pd.read_csv(file_path)

# mirai

file_path = r"data/Mirai-greip_flood.csv"

Mirai_data = pd.read_csv(file_path)

# XSS

file_path = r"data/XSS.csv"

XSS_data = pd.read_csv(file_path)

# Command Injection

file_path = r"data/CommandInjection.csv"

CI_data = pd.read_csv(file_path)

# Backdoor Malware

file_path = r"data/Backdoor_Malware.csv"

BM_data = pd.read_csv(file_path)

# Step 3: Check Packaging

## Brute Force

print(BF_data.info())

print(BF_data.head())


## Benign

print(BN_data.info())

print(BN_data.head())

## Mirai

Mirai_data.info()

Mirai_data.head()

## XSS

XSS_data.info()

XSS_data.head()

## Command Injection

CI_data.info()

CI_data.head()

## Backdoor Malware

BM_data.info()

BM_data.head()

### Results

- keep all communication details
    - `stream`, session unique identifier
    - `user_agent`, flag suspicious devices interactign with cameras
    - **important** `src_mac` seems to be the source device name. this should be kept if I want to focus on a few select cameras or brands 
    - `dst_mac` seems to be the attackers device name. should be kept for analysis
    - `dst_ip`, `dst_port`
- keep `time_since_previously_displayed_frame`: The time since the last similar activity was noticed
- keep `inter_arrival_time` - Sudden changes in timing between packets might suggest an anomaly, like brute force attempts or unauthorized streaming
- data details
    - `eth_size`
    - **important** `payload_length`
            - May highlight abnormal amounts of data being transferred from the camera, like someone streaming the video feed without authorization 
- reliability
    - `l4_tcp` if the data is being sent using a reliable communication method
    - `l4_udp` if the data is being sent using a quicker, less reliable method
- security 
    - `payload_entropy` A way to measure how random or messy the data inside looks—it helps detect suspicious activity
        - High entropy might suggest encrypted or compressed data, which could mean unauthorized parties are exfiltrating sensitive data
    - `user_agent` - May identify unusual or fake devices attempting to interact with cameras
- DNS and HTTPS info
    - `dns_server`, `dns_len_qry`: Look for unusual DNS activity, such as queries to strange domains that don't belong to your system
    - `http_request_method`, `http_host`, `http_uri`: Unusual HTTP requests (e.g., POST when only GET is expected) or access to unauthorized hosts may signal a breach
    - `http_response_code`: Repeated response codes like 401 Unauthorized could indicate unauthorized access attempts
- streaming
    - `stream_jitter`, High jitter could indicate unauthorized streaming or unstable connections
    - `stream_count`, `src_ip_count`, `channel_count`: Spikes in counts could correlate to brute force login attempts, scans, or bulk data transmission


# Step 4: Clean Data

## Drop Unecessary Variables

columns_to_keep = [
    'stream', 'user_agent', 'src_mac', 'dst_mac', 'dst_ip', 'dst_port', 
    'time_since_previously_displayed_frame', 'inter_arrival_time', 'eth_size', 
    'payload_length', 'l4_tcp', 'l4_udp', 'payload_entropy', 'user_agent', 
    'dns_server', 'dns_len_qry', 'http_request_method', 
    'http_host', 'http_uri', 'http_response_code', 'stream_jitter_10_sum',
    'stream_jitter_10_mean', 'stream_jitter_10_var', 'stream_10_count',
    'stream_10_mean', 'stream_10_var',  'src_ip_10_count', 'src_ip_10_mean', 
    'src_ip_10_var', 'channel_10_count', 'channel_10_mean', 'channel_10_var'
]

BF_data = BF_data[columns_to_keep]
BN_data = BN_data[columns_to_keep]
Mirai_data = Mirai_data[columns_to_keep]
XSS_data = XSS_data[columns_to_keep]
CI_data = CI_data[columns_to_keep]
BM_data = BM_data[columns_to_keep]


### drop any NAs

BF_data = BF_data.dropna()
BN_data = BN_data.dropna()
Mirai_data = Mirai_data.dropna()
XSS_data = XSS_data.dropna()
CI_data = CI_data.dropna()
BM_data = BM_data.dropna()


## Focus on Specific Brands/Cameras

# print how many unique devices there are
print(BF_data['src_mac'].nunique())
print(BF_data['src_mac'].unique())

# make new dfs only contianing camera devices being attacked/accessed
camera_BF_data = BF_data[BF_data['src_mac'].str.contains('camera|cam|show', case=False, na=False)]
camera_BN_data = BN_data[BN_data['src_mac'].str.contains('camera|cam|show', case=False, na=False)]
camera_CI_data = CI_data[CI_data['src_mac'].str.contains('camera|cam|show', case=False, na=False)]
camera_Mirai_data = Mirai_data[Mirai_data['src_mac'].str.contains('camera|cam|show', case=False, na=False)]
camera_XSS_data = XSS_data[XSS_data['src_mac'].str.contains('camera|cam|show', case=False, na=False)]
camera_BM_data = BM_data[BM_data['src_mac'].str.contains('camera|cam|show', case=False, na=False)]

# number of unique cameras
print(camera_BF_data['src_mac'].nunique())
print(camera_BN_data['src_mac'].nunique())
print(camera_CI_data['src_mac'].nunique())
print(camera_Mirai_data['src_mac'].nunique())
print(camera_XSS_data['src_mac'].nunique())
print(camera_BM_data['src_mac'].nunique())



# common devices
common_src_mac = set(camera_BF_data['src_mac']).intersection(
    camera_BN_data['src_mac'],
    camera_CI_data['src_mac'],
    camera_Mirai_data['src_mac'],
    camera_XSS_data['src_mac'],
    camera_BM_data['src_mac']
)

camera_BF_data = camera_BF_data[camera_BF_data['src_mac'].isin(common_src_mac)]
camera_BN_data = camera_BN_data[camera_BN_data['src_mac'].isin(common_src_mac)]
camera_CI_data = camera_CI_data[camera_CI_data['src_mac'].isin(common_src_mac)]
camera_Mirai_data = camera_Mirai_data[camera_Mirai_data['src_mac'].isin(common_src_mac)]
camera_XSS_data = camera_XSS_data[camera_XSS_data['src_mac'].isin(common_src_mac)]
camera_BM_data = camera_BM_data[camera_BM_data['src_mac'].isin(common_src_mac)]


print("Common src_mac values:", common_src_mac)


# number of unique cameras
print(camera_BF_data['src_mac'].nunique())
print(camera_BN_data['src_mac'].nunique())
print(camera_CI_data['src_mac'].nunique())
print(camera_Mirai_data['src_mac'].nunique())
print(camera_XSS_data['src_mac'].nunique())
print(camera_BM_data['src_mac'].nunique())

print(common_src_mac)

## Combine Data

camera_BF_data['traffic_type'] = 'BF'
camera_BN_data['traffic_type'] = 'BN'
camera_CI_data['traffic_type'] = 'CI'
camera_Mirai_data['traffic_type'] = 'Mirai'
camera_XSS_data['traffic_type'] = 'XSS'
camera_BM_data['traffic_type'] = 'BM'


combined_data = pd.concat([camera_BF_data, camera_BN_data, camera_CI_data, camera_Mirai_data, camera_XSS_data, camera_BM_data], ignore_index=True)

print(combined_data.info())
# should be 76908 data points 

# numper of instances of each traffic type

traffic_type_counts = combined_data['traffic_type'].value_counts()
print(traffic_type_counts)

## Clean Data Points

combined_data['src_mac'] = combined_data['src_mac'].replace('DCS8000LHA1 D-Link Mini Camera', 'D-Link Mini Camera')

combined_data.sample(10)

combined_data.to_csv('data/CIC_Camera_data.csv', index=False)

## Results


- I will be focusing on the cameras: {'Yi Indoor Camera', 'Nest Indoor Camera', 'AMCREST WiFi Camera', 'Home Eye Camera', 'Rbcior Camera', 'DCS8000LHA1 D-Link Mini Camera', 'Netatmo Camera', 'Arlo Q Indoor Camera', 'TP-Link Tapo Camera', 'HeimVision Smart WiFi Camera', 'Wyze Camera', 'Amazon Echo Show'}
- count of different traffic or attack types
    - BN       35529
    - BF       20552
    - CI        6962
    - XSS       6612
    - BM        6138
    - Mirai     1115
- 76908 total points
