==> 02_InternetDB_Data.ipynb <==
# Accessing InternetDB and Combining with CIC Data

## Overview: Connect to DB, Scrape, Initial Observations, Cleaning dataset, and Combining


---
## Project Workflow:
1. **Step 1: Install and Import Necessary Libraries**
2. **Step 2: Load CIC Dataset**
3. **Step 3: InternetDB Scraping Function**
    - create function
    - test function, see relevant features
4. **Step 4: Create New IP Vulnurability DF**
    - columns from previous relevant features
    - cycle through CIC unique IPs and populate columns with IDB Data



# Step 1: Install and Import Necessary Libraries


import requests
import pandas as pd

# Step 2: Load CIC Dataset


file_path = r"data/CIC_Camera_data.csv"

CIC_data = pd.read_csv(file_path)

CIC_data.sample(10)

# Step 3: InternetDB Scraping Function

# AI-------------
def fetch_internetdb_data(ip):
    url = f"https://internetdb.shodan.io/{ip}"
    response = requests.get(url)
    if response.status_code == 200:
        return response.json()
    else:
        return None

# test function
result = fetch_internetdb_data("35.186.43.132")
print(result)

result = fetch_internetdb_data("47.88.56.147")
print(result)

## Results
- relevant features: open ports, tags, vulnerabilities

# Step 4: Create New IP Vulnurability DF

IP_data = pd.DataFrame(columns=['IP', 'open_ports', 'tags', 'vulns', 'camera_model'])

IP_data

unique_ips = CIC_data['dst_ip'].unique().tolist()

# AI----------------------------
for ip in unique_ips:
    data = fetch_internetdb_data(ip)
    if data:
        camera_model = CIC_data[CIC_data['dst_ip'] == ip]['src_mac'].values[0]
        new_row = pd.DataFrame([{
            'IP': ip,
            'open_ports': data.get('ports', []),
            'tags': data.get('tags', []),
            'vulns': data.get('vulns', []),
            'camera_model': camera_model
        }])
        IP_data = pd.concat([IP_data, new_row], ignore_index=True)


IP_data.sample(10)

IP_data.info()

IP_data.to_csv('data/IP_data.csv', index=False)
