==> 07_Device_Info.ipynb <==
# Using InternetDB

## Overview: Gather info on unique IP adresses


---
## Project Workflow:
1. **Step 1: Install and Import Necessary Libraries**
2. **Step 2: Load Data**
3. **Step 3: Analyze Device IPs**



# Step 1: Install and Import Necessary Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy.stats import pearsonr
import ast
from collections import Counter

# Step 2: Load Data - Only 2-Model System

model_info = pd.read_csv("data/twoM_devices.csv")

model_info = model_info[model_info['Metric'] == 'Recall']
model_info

model_info = model_info.groupby('src_mac')['Value'].mean().reset_index()

model_info.columns = ['src_mac', 'Recall']
model_info = model_info.rename(columns={'src_mac': 'camera_model'})

model_info

model_info.info()

## load IP Data

IP_data = pd.read_csv("data/IP_data.csv")

IP_data.info()

# fixing the data types
IP_data['tags'] = IP_data['tags'].apply(ast.literal_eval)
IP_data['open_ports'] = IP_data['open_ports'].apply(ast.literal_eval)
IP_data['vulns'] = IP_data['vulns'].apply(ast.literal_eval)

IP_data[IP_data['camera_model'] == 'Rbcior Camera'].sample(5)



# Step 3: Analyze Device IPs

## Merge dfs

IP_data = pd.merge(IP_data, model_info, on='camera_model', how='left')

IP_data.info()

IP_data.sample(5)

## Analysis

### Does number of Open Ports affect recall?

def count_OPs(ops):
    return len(ops)

IP_data['num_OPs'] = IP_data['open_ports'].apply(count_OPs)

IP_data[['camera_model', 'num_OPs']].sample(5)

correlation, p_value = pearsonr(IP_data['num_OPs'], IP_data['Recall'])

print(f"Pearson coefficient: {correlation}")
print(f"p-value: {p_value}")

!!! no significant negative correlation with number of open ports with recall score

port_counts = Counter(port for port_list in IP_data['open_ports'] for port in port_list)

print(port_counts)

for port in port_counts.keys():
    IP_data[f'port_{port}'] = IP_data['open_ports'].apply(lambda ports: 1 if port in ports else 0)

for port in port_counts.keys():
    correlation, p_value = pearsonr(IP_data[f'port_{port}'], IP_data['Recall'])
    if p_value < 0.05:
        print(f"Port {port} is significant. p-value: {p_value}, Pearson coefficient: {correlation}")


#### Results
- Port 10001
    - P2P live streaming on cameras like Wyze
- port 8080
    - port 80 alternative
    - http traffic
    - might suggest web server traffic, proxy server usage, or testing and development activities
- port 32100
    - common for p2p communication for many IoT cameras

- P2P is a way for computers and devices to communicate directly with each other without needing a central server
    - share data directly
    - file sharing


### Does num Vulnurabilities affect recall?

def count_vulns(vulns):
    return len(vulns)

IP_data['num_vulns'] = IP_data['vulns'].apply(count_vulns)

IP_data[['camera_model', 'num_vulns']].sample(5)

# Perform Pearson correlation test
correlation, p_value = pearsonr(IP_data['num_vulns'], IP_data['Recall'])

# Print the results
print(f"Pearson coefficient: {correlation}")
print(f"p-value: {p_value}")


#### Results
significant negative correlation with number of vulnurabilities with recall score

### Investigate specific vulns on each camera model and what they correspond to



vulns_count = Counter(vuln for vulns_list in IP_data['vulns'] for vuln in vulns_list)

print(vulns_count)

for vuln in vulns_count.keys():
    IP_data[f'vuln_{vuln}'] = IP_data['vulns'].apply(lambda vulns: 1 if vuln in vulns else 0)

for vuln in vulns_count.keys():
    correlation, p_value = pearsonr(IP_data[f'vuln_{vuln}'], IP_data['Recall'])
    if p_value < 0.05:
        print(f"Vuln {vuln} is significant. p-value: {p_value}")


#### Results

| Year | Significant Vulnerabilities |
|------|----------------|
| 2016 | 8743, 4975, 4979, 5387, 1546, 8740 |
| 2017 | 9788, 3169, 7679, 15710, 3167, 15715 |
| 2018 | 17199, 1312, 1333, 1283, 17189, 11763 |
| 2019 | 0196, 0211, 10082, 0217, 10098, 10092, 0220, 17567 |
| 2020 | 35452, 1927, 1934, 13938, 11985 |
| 2021 | 26690, 33193, 44224, 26691 |
| 2022 | 36760, 23943, 26377 |
| 2023 | 45802, 25690 |
| 2024 | 38474, 38476, 27316, 38477 |

### Are any tags associated with lower recalls?

tag_counts = Counter(tag for tags_list in IP_data['tags'] for tag in tags_list)

print(tag_counts)

for tag in tag_counts.keys():
    IP_data[f'tag_{tag}'] = IP_data['tags'].apply(lambda tags: 1 if tag in tags else 0)


IP_data.info()

# Ai ----------------------------------------
print("Cloud")
correlation, p_value = pearsonr(IP_data['tag_cloud'], IP_data['Recall'])

print(f"Pearson correlation coefficient: {correlation}")
print(f"P-value: {p_value}")

print("self_signed")
correlation, p_value = pearsonr(IP_data['tag_self-signed'], IP_data['Recall'])

print(f"Pearson correlation coefficient: {correlation}")
print(f"P-value: {p_value}")

print("cdn")
correlation, p_value = pearsonr(IP_data['tag_cdn'], IP_data['Recall'])

print(f"Pearson correlation coefficient: {correlation}")
print(f"P-value: {p_value}")

print("eol_product")
correlation, p_value = pearsonr(IP_data['tag_eol-product'], IP_data['Recall'])

print(f"Pearson correlation coefficient: {correlation}")
print(f"P-value: {p_value}")

print("starttls")
correlation, p_value = pearsonr(IP_data['tag_starttls'], IP_data['Recall'])

print(f"Pearson correlation coefficient: {correlation}")
print(f"P-value: {p_value}")


print("database")
correlation, p_value = pearsonr(IP_data['tag_database'], IP_data['Recall'])

print(f"Pearson correlation coefficient: {correlation}")
print(f"P-value: {p_value}")


#### Results
- eol product is close to 0.05 alpha but not less, none are significantly correlated at 0.05 level
- eol product is end of life product
    - no longer recieve updates from vendor
        - unpatched vulns
